refactor(zaznamnik): split tab into PhArchView + VtTestGroup, narrow memo deps
