+++ src/components/SubjectFileDrawer/index.tsx
@@ -23,16 +23,18 @@ export function SubjectFileDrawer({ lesson, isOpen, onClose }: { lesson: BlockLe
     const classmatesCount = useAppStore(s => lesson?.courseCode ? s.classmates[lesson.courseCode]?.length : undefined);
     const zaznamnikData = useAppStore(s => lesson?.courseCode ? s.zaznamnik?.[lesson.courseCode] : undefined);
 
+    const phSections = zaznamnikData?.ph.sections;
+    const vtTests = zaznamnikData?.vt.tests;
     const tabCounts = useMemo(() => {
-        const phCount = zaznamnikData?.ph.sections.flatMap(s => s.arches).filter(a => !a.empty).length ?? 0;
-        const vtCount = zaznamnikData?.vt.tests.length ?? 0;
+        const phCount = phSections?.reduce((n, s) => n + s.arches.filter(a => !a.empty).length, 0) ?? 0;
+        const vtCount = vtTests?.length ?? 0;
         return {
             files: state.files?.reduce((acc, f) => acc + f.files.length, 0) || 0,
             classmates: classmatesCount || 0,
             // undefined = not yet loaded; 0 = loaded but empty; >0 = has data
             zaznamnik: zaznamnikData !== undefined ? phCount + vtCount : undefined,
         };
-    }, [state.files, classmatesCount, zaznamnikData]);
+    }, [state.files, classmatesCount, zaznamnikData, phSections, vtTests]);
 
     useEffect(() => {
         if (isOpen && state.files?.length) {
